Move shared Classes to library
