+++ io.ebu.eis.shared/io.ebu.eis.shared.csproj
@@ -1,0 +1,101 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{88FBE6F3-7FC9-4940-8403-BBCF93BE6EDC}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <AppDesignerFolder>Properties</AppDesignerFolder>
+    <RootNamespace>io.ebu.eis.shared</RootNamespace>
+    <AssemblyName>io.ebu.eis.shared</AssemblyName>
+    <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
+    <FileAlignment>512</FileAlignment>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug\</OutputPath>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
+    <Reference Include="System" />
+    <Reference Include="System.Configuration" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Drawing" />
+    <Reference Include="System.Runtime.Serialization" />
+    <Reference Include="System.Windows" />
+    <Reference Include="System.Xaml" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Xml" />
+    <Reference Include="WindowsBase" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="JsonSerializer.cs" />
+    <Compile Include="ManagerImageReference.cs" />
+    <Compile Include="ManagerTemplateField.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Utils\AspectRatioConverter.cs" />
+    <Compile Include="Utils\HeightAspectRatioConverter.cs" />
+    <Compile Include="Utils\RegistryHelper.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\io.ebu.eis.canvasgenerator\io.ebu.eis.canvasgenerator.csproj">
+      <Project>{490ca7a2-b578-490f-9561-4892d96969d4}</Project>
+      <Name>io.ebu.eis.canvasgenerator</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\io.ebu.eis.data.file\io.ebu.eis.data.file.csproj">
+      <Project>{06e8f7af-da76-4f90-bcc6-b43aabd4414c}</Project>
+      <Name>io.ebu.eis.data.file</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\io.ebu.eis.data.ftp\io.ebu.eis.data.ftp.csproj">
+      <Project>{4d97fb11-b83b-47df-8fe2-70a94a10385d}</Project>
+      <Name>io.ebu.eis.data.ftp</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\io.ebu.eis.data.s3\io.ebu.eis.data.s3.csproj">
+      <Project>{afbc1f4d-b0ad-4ac8-8d0d-ac2b7298c4bd}</Project>
+      <Name>io.ebu.eis.data.s3</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\io.ebu.eis.datastructures\io.ebu.eis.datastructures.csproj">
+      <Project>{7ce61fc1-d140-4eb3-b2a0-efd395f8d1db}</Project>
+      <Name>io.ebu.eis.datastructures</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\io.ebu.eis.http\io.ebu.eis.http.csproj">
+      <Project>{3e251eb6-130a-49e1-80b7-ba98879162d1}</Project>
+      <Name>io.ebu.eis.http</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\io.ebu.eis.mq\io.ebu.eis.mq.csproj">
+      <Project>{36c58a7c-8fc5-4290-bcd5-8b56e2802497}</Project>
+      <Name>io.ebu.eis.mq</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\io.ebu.eis.stomp\io.ebu.eis.stomp.csproj">
+      <Project>{02a3fd2c-6dc9-434f-a28a-cb8aff8d50ea}</Project>
+      <Name>io.ebu.eis.stomp</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
+       Other similar extension points exist, see Microsoft.Common.targets.
+  <Target Name="BeforeBuild">
+  </Target>
+  <Target Name="AfterBuild">
+  </Target>
+  -->
+</Project>
